feat: take input data from tree sequences
* fix unrelated lint
