+++ src/test.rs
@@ -199,7 +199,7 @@ chr0	1	.	G	A	.	.	.	GT	/0	/1	/1	/0	/1	/1	/0	/0	/.	/.	/0	/0	/1	/1	/1	/1	/0	/."#
 
         #[test]
         fn load_vcf() {
-            let (all_alleles, allele_counts) = counts_from_vcf(make_vcf(), 1);
+            let (_all_alleles, allele_counts) = counts_from_vcf(make_vcf(), 1);
             let mut iter = allele_counts.iter();
             let counts_0 = iter.next().unwrap();
             let counts_1 = iter.next().unwrap();
